+++ docs/CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 1.2.0  (March 10, 2021)
 --------------------------
 - Enh: Replace depreciated MarkdownField widget with new RichTextField widget
+- Enh: Add an option to show legal pages in full screen after account creation ([#23](https://github.com/humhub-contrib/legal/issues/23))
+- Enh: Add missing translations (automatic from Google translation)
 
 
 1.1.4  (February 17, 2021)